+++ app/domain/game/components/AttendanceBanner/index.tsx
