+++ app/domain/game/components/AttendanceBanner/style.css.ts
