+++ app/domain/game/components/AttendancePolicy/index.tsx
