+++ app/domain/game/components/AttendancePolicy/style.css.ts
